-  Auth with password-less method -  Commander to create mail instance -  Create limiter for guest and auth -  Create notify method in model User

+++ app/commands/mail.js
@@ -1,51 +1,24 @@
 const program = require('commander');
 const fs = require('fs');
 const path = require('path');
+const mailTemplate = require('../../resources/templates/command/mail');
 
 program
-  .command('create <classname>')
-  .description('Create a new mail class')
-  .action((classname) => {
-    const filePath = path.join(__dirname, '../mails' ,`${classname}.js`);
-    const fileContent = 
-`class ${classname} {
-    constructor() {
-    }
+    .command('create <classname>')
+    .description('Create a new mail class')
+    .action((classname) => {
+        const filePath = path.join(__dirname, '../mails', `${classname}.js`);
+        const fileContent = mailTemplate(classname);
 
-    from() {
-        return '';
-    }
-
-    subject() {
-        return '';
-    }
-
-    greeting() {
-        return '';
-    }
-
-    content() {
-        return \`\`;
-    }
-
-    button() {
-        return {
-            text: '',
-            link: ''
+        if (fs.existsSync(filePath)) {
+            console.error(`File ${classname} already exists`);
+            return;
         }
-    }
 
-    footer() {
-        return '';
-    }
-}
-
-module.exports = ${classname};`;
-
-    fs.writeFile(filePath, fileContent, (err) => {
-      if (err) throw err;
-      console.log(`Mail class ${classname}.js created successfully`);
+        fs.writeFile(filePath, fileContent, (err) => {
+            if (err) throw err;
+            console.log(`Mail class for ${classname}.js created successfully`);
+        });
     });
-  });
 
 program.parse(process.argv);
